--- .github/actions/javascript/getArtefactInfo/action.yml
@@ -1,19 +1,0 @@
-name: "Get artefact info"
-description: "Gets the ID and workflow ID about a specific artefact. By default artifacts are only available in the same workflow. This action can be used to get the info to download an artefact from a different workflow."
-inputs:
-  GITHUB_TOKEN:
-    description: Auth token for New Expensify Github; necessary for accessing Octokit.
-    required: true
-  ARTEFACT_NAME:
-    description: Name of the artefact to get infos about (e.g. to use for downloading that artefact)
-    required: true
-outputs:
-  ARTEFACT_FOUND:
-    description: Whether the artefact was found
-  ARTEFACT_ID:
-    description: The ID of the artefact
-  ARTEFACT_WORKFLOW_ID:
-    description: The ID of the workflow that produced the artefact
-runs:
-  using: "node20"
-  main: "index.js"
