+++ .github/actions/javascript/getArtifactInfo/action.yml
@@ -1,0 +1,19 @@
+name: "Get artifact info"
+description: "Gets the ID and workflow ID about a specific artifact. By default artifacts are only available in the same workflow. This action can be used to get the info to download an artifact from a different workflow."
+inputs:
+  GITHUB_TOKEN:
+    description: Auth token for New Expensify Github; necessary for accessing Octokit.
+    required: true
+  ARTIFACT_NAME:
+    description: Name of the artifact to get infos about (e.g. to use for downloading that artifact)
+    required: true
+outputs:
+  ARTIFACT_FOUND:
+    description: Whether the artifact was found
+  ARTIFACT_ID:
+    description: The ID of the artifact
+  ARTIFACT_WORKFLOW_ID:
+    description: The ID of the workflow that produced the artifact
+runs:
+  using: "node20"
+  main: "index.js"
